fold recording order regression test into lifecycle tests
